クリーンして作成中
MemberInfoを削除
再構築します
