+++ projects/LabMemberManager/backup/MemberInfoBean.java
